fix: add email visibility option in simple checkout
This commit adds the 'showEmail' option to the simple checkout feature. This option allows for the
field to be made visible in the checkout if required. It also changes some settings initialization
to use null coalescing operator for more concise code and updates some additional related files to
reflect this change. The new dependency 'quiqqer/order-guestorder' has been added to composer.json.
A new locale entry for email has been added for both German and English.