fix: add 'orderInvalid' event and correct email display
In `SimpleCheckout.js`, an 'orderInvalid' event is now fired in multiple conditions where the order
form does not pass the validity checks. This will allow other components to listen for this event
and take appropriate actions.

In `CheckoutDelivery.php`, logic has been added to handle the display of the email field during
checkout. The email field will now be hidden if the user is logged in and does not have an email
associated with their account. Therefore, the field 'showEmail' from Checkout is now altered
depending on the user's session.