feat: improve checkout process by handling mandatory links efficiently
In this commit, the checkout process in QUI has been improved.

Firstly, CSS for displaying mandatory links in the checkout form has been added (displaying these
links as a column of checkboxes). This provides a convenient overview of the conditions to be
accepted before checkout.

Secondly, the html template for the checkout form has been updated to display the checkboxes
depending on the type of mandatory links setting (multiple_checkboxes or single_checkbox). The name
and description for each checkbox are fetched dynamically. This increases the customizability of
the form.

Thirdly, PHP code was updated in the Checkout class to generate the list of checkboxes dynamically.
Different conditions (terms and conditions, privacy policy, revocation) each have their own
checkbox. Images are also assigned accordingly. This increases readability and eases maintenance.

Finally, a general refactoring was done to reduce redundancy and improve the readability of the
code. The repetitive code for terms and condition setup was removed and the functionality was
optimized.

Overall, these changes make the code more efficient, cleaner, and easier to maintain. The checkout
form also becomes more user-friendly.