fix: cast product id to integer and validate order hash
In `newOrderInProcess.php`, the product id is now cast to an integer for better reliability and
consistency in its usage.

In `SimpleCheckout.js`, added a validation to ensure the order hash is a valid UUID before using
it. Additional check has been added so the window location hash would not be overwritten if it
already exists.